fix(atp): widen statvfs blocks_available to u64 before frsize multiply
nix on this platform returns u32 for blocks_available while fragment_size
is u64; the raw checked_mul failed E0308 and broke compilation of the
atp_multi_peer_integration target, red-gating the whole suite that the
0.4.10 release-readiness sweep needs. u64::from is lossless.

+++ tests/atp/forensics.rs
@@ -581,7 +581,7 @@ fn capture_system_state(output_dir: &Path) -> SystemState {
 #[cfg(unix)]
 fn available_disk_space(path: &Path) -> Option<u64> {
     let stat = nix::sys::statvfs::statvfs(path).ok()?;
-    stat.blocks_available().checked_mul(stat.fragment_size())
+    u64::from(stat.blocks_available()).checked_mul(stat.fragment_size())
 }
 
 #[cfg(not(unix))]
